perf(package/server): 将单index.ts初步拆分成多文件结构；分离库单元和运行文件
