+++ packages/server/src/types.ts
@@ -1,0 +1,8 @@
+export interface ServerOptions {
+	port?: number;
+}
+
+export interface ServerInstance {
+	start(): Promise<void>;
+	stop(): Promise<void>;
+}
