+++ packages/server/tsup.config.ts
@@ -1,8 +1,8 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: ['src/index.ts'],
-  format: ['esm', 'cjs'],
+  entry: ["src/index.ts", "src/cli.ts"],
+  format: ["esm", "cjs"],
   dts: true,
   splitting: false,
   sourcemap: true,
